multi-arfcn, trx: add polyphase channelizer and synthesis filter
Polyphase channelizer allows for the use of 'M' independent
channels over a single physical device. In contrast to a
direct frequency translate and filter approach, the polyphase
implementation shifts processing from the high sampling rate
side and allows operation at the lower, more efficient
channel rate.

Multirate channelizer design consists of two stages. From the
perspective of the receive side, the first stage consists of
feeding the receive signal - sampled at 'M' times the channel
spacing - through a 'M' path polyphase filter. The output of
the filter consists of all channels aliased down to baseband.
Phase alignment is performed to separate the alised signals;
an FFT is used for efficient implementation.

In the second stage, the 'M' separated channels at the output
of the FFT are selectively downsampled from channel rate to
the GSM symbol rate. Another polyphase filterbank is used for
this purpose.

The transmit side "synthesis" filter combines 'M' channels
into a single output channel and is the inverse of the
aforementioned channelization process.

Signed-off-by: Thomas Tsou <ttsou@vt.edu>