multi-arfcn, uhd: output floating point directly from uhd
The transceiver core operates on floating point so we can
find some savings in performing the conversion within UHD.
The conversion is then done on the kernel-to-userspace
copy, which needs to occur anyways. We take on a slight
I/O penalty from the larger sample sizes, but this is
negated by other performance benefits.

Signed-off-by: Thomas Tsou <ttsou@vt.edu>