multi-arfcn, core: setup gsm core for handle multi-carrier
A single Transceiver manager manages a list of pointers
to 'M' ARFCN managers. All accesses to an ARFCN manager,
or 'radio', require specification of the channel number
ranging from 0 to 'M-1'. Channel number is related to,
but may not correspond to ARFCN number in that the chan-
nel number depends on the particular channelizer setting.

Signed-off-by: Thomas Tsou <ttsou@vt.edu>