multi-arfcn, sigproc: set aligned memory allocation on 32-bit machines
Wel need 16-byte alingment for SSE aligned loads. GCC will guarantee
that on 64-bit architectures, but only 8-byte alignment on 32 machines.
In any case, use memalign to explicity request 16-byte alignment.

This primarily affects filter taps, which are assumed to be always
be read from index 0 of the allocated vector.

Signed-off-by: Thomas Tsou <ttsou@vt.edu>