multi-arfcn, trx: remove limitation on transceiver sample rate
There was a previous assumption that the transceiver would
operate at the GSM symbol rate and the channelizer would run
at a higher rate (generally 400 ksps). With two samples-per-
symbol in the transceiver, the channelizer needs to run at the
lower of the two rates. To handle both cases, we need to be
explicit when specifying filter and buffer construction in the
base class so that the lengths are setup correctly.

Signed-off-by: Thomas Tsou <ttsou@vt.edu>