multi-arfcn, usrp: disable USRP1 with multi-arfcn
Since we're now outputing floating point values out
of UHD, the interface is changed such that the
integer values from USRP1 are no longer supported.
Also, use of the USRP1 with GNU Radio driver at sample
rates beyond 270.8333 ksps and 400 ksps are unknown
at this time.

This does not preclude the use of USRP1 with multi-
arfcn support, which may be added at a later time.
Disable the USRP1 build until that support is added.

Signed-off-by: Thomas Tsou <ttsou@vt.edu>