multi-arfcn, usrp1: normalize sample amplitude and add device offsets
We need to reduce each transceiver output amplitude to prevent
splatter on the device transmitter input. The scaling is not
efficient; for now, so be it.

Also add device specific timing offsets to the paramter table. UHD
devices other than the USRP1 are mostly consistent on tx/rx chain
offsets, but the USRP1 is far out enough to require it's own
settings.

Signed-off-by: Thomas Tsou <ttsou@vt.edu>