transceiver: mcbts: add third resampling stage
This resampling stage performs rate conversion of the device sample
rate to a multiple of the channelizer rate (400 kHz). The other
resampler stages are channelization to 400 kHz and the GSM specific
conversion a multiple of the 270.833 kHz symbol rate.

Benefits of this patch are the use of sample rates unsupported by
the device as well as selection of rates with minimal passband
distortion and out-out-band emissions. Desireable rates for
supported devices use decimation rates that are multiples of 4,
which causes both half band filters to be used in the FPGA. This
reduces CIC distortion and aliasing out-of-band.

Signed-off-by: Thomas Tsou <tom@tsou.cc>