mcbts: resolve potential aliasing on large downsampling ratios
The decimation factor was being ignored during filter construction,
which would cause the resampling filter to be overly wide. The
consequence is aliasing that may increase as the input signal
bandwidth increases. The preferred approach, which this patch
changes, is to bandlimit the signal rather than allowing aliasing
to occur.

Signed-off-by: Thomas Tsou <tom@tsou.cc>